Add blank VSchema for new keyspaces
This change adds a blank VSchema for a newly created keyspace. As
keyspaces can't be used until they have a vschema this allows an
unsharded keyspace to be used without requireing an explicit vschema
to be loaded for it
